# nodejs-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Api Documentation"
  version: "1.0"
  title: "Api Documentation"
  termsOfService: "urn:tos"
  contact: {}
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0"
host: "localhost:8080"
basePath: "/"
...
paths:
...
  /product:
    get:
      tags:
      - "controller"
      summary: "product"
      operationId: "productUsingGET"
      produces:
      - "*/*"
      parameters:
      - name: "Authorization"
        in: "header"
        description: "Authorization"
        required: true
        type: "string"
      - name: "Product"
        in: "query"
        description: "Product"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "OK"
          schema:
            type: "object"
        "401":
          description: "Unauthorized"
        "403":
          description: "Forbidden"
        "404":
          description: "Not Found"
      x-swagger-router-controller: "Controller"
...